requirements.txt
playwright>=1.42.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.31.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0
